refactor: 💡 [EXL-66] split files and fix position of exlint logo in auth page
split files and fix position of exlint logo in auth page
