+++ apps/frontend/src/components/containers/Auth/Auth.module.scss
@@ -1,137 +1,4 @@
-@use 'sass:map';
-
-@import '../../../styles/variables.scss';
-
 .main {
 	display: flex;
 	font-family: Poppins, sans-serif;
-
-	.leftSideContainer {
-		position: relative;
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		justify-content: center;
-		width: 50%;
-		padding-inline: map.get($sizes, spacing-xxl-2);
-		background-image: url('../../../assets/images/auth-background.png');
-		background-repeat: no-repeat;
-		background-position: center;
-		background-size: cover;
-
-		&__brandLogo {
-			position: absolute;
-			top: map.get($sizes, spacing-xxl);
-			inset-inline-start: map.get($sizes, spacing-l);
-			width: 168px;
-			object-fit: contain;
-		}
-
-		&__header {
-			margin-bottom: map.get($sizes, spacing-xl);
-			font-size: 4.5rem;
-			font-weight: 700;
-			line-height: 58.75px;
-			text-align: center;
-			background: linear-gradient(250deg, #dc6794 6.97%, #9747ff 93.64%);
-			background-clip: text;
-			-webkit-text-fill-color: transparent;
-		}
-
-		&__subHeader {
-			font-size: 3.6rem;
-			font-weight: 400;
-			color: map.get($colors, greys-gainsboro);
-			text-align: center;
-		}
-	}
-
-	.rightSideContainer {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		justify-content: center;
-		width: 50%;
-
-		&__header {
-			font-size: 4rem;
-			font-weight: 700;
-			color: map.get($colors, purples-purple);
-		}
-
-		&__subHeader {
-			margin-bottom: map.get($sizes, spacing-xxl-2);
-			font-size: 3.5rem;
-			font-weight: 400;
-			color: map.get($colors, purples-purple);
-		}
-
-		.rightSideExternalAuthContainer {
-			display: flex;
-			align-items: center;
-			width: 357px;
-			height: 67px;
-			padding: map.get($sizes, spacing-m) map.get($sizes, spacing-xl);
-			text-decoration: none;
-			border-radius: 10px;
-
-			&:not(:last-child) {
-				margin-bottom: map.get($sizes, spacing-l);
-			}
-
-			&--github {
-				background-color: map.get($colors, blacks-yankees-black-blue);
-				border: 1.5px solid transparent;
-			}
-
-			&--google {
-				background-color: map.get($colors, whites-white);
-				border: 1.5px solid map.get($colors, greys-platinum);
-			}
-
-			&__brandLogo {
-				height: 43.1px;
-				margin-inline-end: map.get($sizes, spacing-xxl);
-				object-fit: contain;
-			}
-
-			&__text {
-				font-size: 1.9rem;
-				font-weight: 500;
-
-				&--github {
-					color: map.get($colors, greys-anti-flash-grey-white);
-				}
-
-				&--google {
-					color: map.get($colors, blacks-dark-gunmetal);
-				}
-			}
-		}
-
-		&__divider {
-			width: 203px;
-			height: 2px;
-			margin-block: map.get($sizes, spacing-xxl-2);
-			background-color: map.get($colors, greys-platinum);
-		}
-
-		&__permissionsDisclaimer {
-			margin-bottom: map.get($sizes, spacing-m);
-			font-size: 1.6rem;
-			font-weight: 300;
-			color: map.get($colors, greys-independence-grey);
-		}
-
-		&__policyText {
-			font-size: 1.6rem;
-			font-weight: 300;
-			color: map.get($colors, greys-independence-grey);
-			text-align: center;
-		}
-
-		&__textLink {
-			color: map.get($colors, greys-independence-grey);
-		}
-	}
 }
