+++ apps/frontend/src/components/containers/Auth/Auth.tsx
@@ -1,16 +1,11 @@
 import React from 'react';
-import { useSearchParams } from 'react-router-dom';
 
 import AuthView from './Auth.view';
 
 interface IProps {}
 
 const Auth: React.FC<IProps> = () => {
-	const [searchParams] = useSearchParams();
-
-	const port = searchParams.get('port');
-
-	return <AuthView port={port} />;
+	return <AuthView />;
 };
 
 Auth.displayName = 'Auth';
