+++ apps/frontend/src/components/containers/Auth/LeftSide/LeftSide.module.scss
@@ -1,0 +1,43 @@
+@use 'sass:map';
+
+@import '../../../../styles/variables.scss';
+
+.container {
+	position: relative;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	width: 50%;
+	padding-inline: map.get($sizes, spacing-xxl-2);
+	background-image: url('../../../../assets/images/auth-background.png');
+	background-repeat: no-repeat;
+	background-position: center;
+	background-size: cover;
+
+	&__brandLogo {
+		position: absolute;
+		top: map.get($sizes, spacing-l);
+		inset-inline-start: map.get($sizes, spacing-l);
+		width: 168px;
+		object-fit: contain;
+	}
+
+	&__header {
+		margin-bottom: map.get($sizes, spacing-xl);
+		font-size: 4.5rem;
+		font-weight: 700;
+		line-height: 58.75px;
+		text-align: center;
+		background: linear-gradient(250deg, #dc6794 6.97%, #9747ff 93.64%);
+		background-clip: text;
+		-webkit-text-fill-color: transparent;
+	}
+
+	&__subHeader {
+		font-size: 3.6rem;
+		font-weight: 400;
+		color: map.get($colors, greys-gainsboro);
+		text-align: center;
+	}
+}
