+++ apps/frontend/src/components/containers/Auth/LeftSide/index.ts
@@ -1,0 +1,3 @@
+import LeftSide from './LeftSide';
+
+export default LeftSide;
