+++ apps/frontend/src/components/containers/Auth/RightSide/index.ts
@@ -1,0 +1,3 @@
+import RightSide from './RightSide';
+
+export default RightSide;
